Feature/4

+++ src/main/java/com/aswemake/api/aswemakeapitask/dto/item/response/ItemCreateResponseDto.java
@@ -13,6 +13,6 @@ import lombok.NoArgsConstructor;
 public class ItemCreateResponseDto {
     private Long id;
     private String name;
-    private int price;
+    private Long price;
     private int stockQuantity;
 }
