+++ src/main/java/com/aswemake/api/aswemakeapitask/dto/item/response/ItemUpdateResponseDto.java
@@ -14,8 +14,8 @@ import lombok.NoArgsConstructor;
 public class ItemUpdateResponseDto {
     private Long id;
     private String name;
-    private int beforePrice; //수정 전 가격
-    private int afterPrice; //수정 후 가격
+    private Long beforePrice; //수정 전 가격
+    private Long afterPrice; //수정 후 가격
     private int stockQuantity;
     private int remainingStockQuantity;   //수정 당시 남은 재고 수량
 }
